refactor : Validation 객체로 책임 이동, 역할 위임
