+++ src/test/java/lotto/step1/CalculatorTest.java
@@ -8,39 +8,46 @@ import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
 
 class CalculatorTest {
-    @Test
-    @DisplayName("문자열로 숫자 하나를 입력 했는지 검증 하는 테스트")
-    void onlyNumberTest() {
-        Calculator calculator = new Calculator("3");
-        Assertions.assertThat(calculator.checkOnlyNumber()).isTrue();
-    }
-    @Test
-    @DisplayName("문자열로 숫자 하나를 입력 했는지 검증 하는 테스트2")
-    void onlyNumberTest2() {
-        Calculator calculator = new Calculator("s");
-        Assertions.assertThat(calculator.checkOnlyNumber()).isFalse();
-    }
+//    @Test
+//    @DisplayName("빈 문자열 이나 null 을 입력했을 때 boolean 값 확인")
+//    void checkEmptyAndNull(){
+//        Calculator calculator = new Calculator(null);
+//        assertThat(calculator.checkEmptyAndNull()).isTrue();
+//    }
+//
+//    @Test
+//    @DisplayName("문자열로 숫자 하나를 입력 했는지 검증 하는 테스트")
+//    void onlyNumberTest() {
+//        Calculator calculator = new Calculator("3");
+//        assertThat(calculator.checkOnlyNumber()).isTrue();
+//    }
+//    @Test
+//    @DisplayName("문자열로 숫자 하나를 입력 했는지 검증 하는 테스트2")
+//    void onlyNumberTest2() {
+//        Calculator calculator = new Calculator("s");
+//        assertThat(calculator.checkOnlyNumber()).isFalse();
+//    }
 
     @Test
     @DisplayName("음수를 포함했을 때 예외처리 하는 테스트")
     void isNegativeNumber() {
         Calculator calculator = new Calculator("-1,2,4");
-        Assertions.assertThatThrownBy(() -> calculator.isNegativeNumber("-1")).isInstanceOf(RuntimeException.class);
+        assertThatThrownBy(() -> calculator.isNegativeNumber("-1")).isInstanceOf(RuntimeException.class);
     }
 
-    @Test
-    @DisplayName("문자열에 커스텀 구분자가 있는지 확인하는 테스트")
-    void checkCustomDelimiter() {
-        Calculator calculator = new Calculator("//;\n1;2;3");
-        assertThat(calculator.checkCustomDelimiter()).isTrue();
-    }
-
-    @Test
-    @DisplayName("문자열에 커스텀 구분자가 있는지 확인하는 테스트2")
-    void checkCustomDelimiter2() {
-        Calculator calculator = new Calculator("//.\n3.4.2");
-        assertThat(calculator.checkCustomDelimiter()).isTrue();
-    }
+//    @Test
+//    @DisplayName("문자열에 커스텀 구분자가 있는지 확인하는 테스트")
+//    void checkCustomDelimiter() {
+//        Calculator calculator = new Calculator("//;\n1;2;3");
+//        assertThat(calculator.checkCustomDelimiter()).isTrue();
+//    }
+//
+//    @Test
+//    @DisplayName("문자열에 커스텀 구분자가 있는지 확인하는 테스트2")
+//    void checkCustomDelimiter2() {
+//        Calculator calculator = new Calculator("//.\n3.4.2");
+//        assertThat(calculator.checkCustomDelimiter()).isTrue();
+//    }
 
     @Test
     @DisplayName("빈 문자열이나 null 일 때 0 반환하는 테스트")
